Documented websocket.upgradeReq
In my attempt to associate a websocket with a specific client I had to override `websocket.handleUpgrade` and call it abstractly inorder to get both the request and the websocket in one place. If I would have known that it was available, I wouldn't have to code so dirty for something I think many people would need.