Re-implement custom pause/resume methods and test
Implemented in order to prevent breakage due to the assumption that
pausing the stream itself will also immediately pause the underlying
socket. This implementation fix will ensure that the underlying socket
is paused/resumed when the stream is paused/resumed, but should do so
without affecting the actual pause and resume methods.