[major] Remove the `flags` argument (#1101)
The `flags` object specifies whether the received data is binary or not
and whether it was masked or not. This object is actually redundant and
therefore removed in this commit as the spec requires client-sent
frames to be masked and server-sent frame to be unmasked.

The `typeof` operator can be used to check whether the received data is
binary or not.