[fix] Prevent type error in Duplex stream implementation on Node >=12
Recent versions of Node have added getters for some of the internal stream
properties and throws this state away when the stream is being ended. This
commit adds a polyfill for the internal properties that are now exposed
through getters, and checks for availability of properties that do not have
a corresponding getter before using them.