Fix close-after-error behaviour in Node < v14
If we're already closing, then we shouldn't move into the new -2 ready
state (i.e. closing-because-the-remote-end-closed). We're already
closing, we shouldn't go back to a state that allows sending another
close frame.

This didn't seem to affect the happy path, or v14, or the existing
tests, but in the new error test on v10 and similar it resulted in a
double close(), and a write-after-end error on the socket itself.