[fix] make test/create-websocket-stream.test.js compatible with domains
Fix a failure in test/create-websocket-stream.test.js if the node:domain
module is loaded (e.g. due to NODE_OPTIONS in environment).

The cause of the failure was that installing an 'uncaughtException'
event handler on process will cause node:domain to prepend its own
handler for the same event, which confused the test.

Fixes: #2124