feat(events): emitting a 'closing' event when the server starts the close handshake
This event helps us control the flow of streaming data to the remote websocket server when a close
handshake is initiated. The 'close' event is normally only triggered after the connection is
already closed, but at this time it is possible that data has been written to the socket, which is
no longer available. When the 'closing' event is now emitted, this means that it is time to stop
sending any data upstream because a close event will follow soon.